feat: Update oauth2-proxy component
This commit updates the oauth2-proxy component by making the following changes:
- Updated imports in component.ts to include additional dependencies
- Removed unused code and interfaces from index.ts
- Added a main entry point in package.json
- Removed unnecessary compiler options from tsconfig.json
