+++ components/oauth2-proxy/package.json
@@ -2,6 +2,7 @@
   "name": "@unmango/pulumi-oauth2-proxy",
   "version": "0.1.0",
   "description": "OAuth2 Proxy Pulumi component",
+  "main": "index.ts",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "build": "tsc --outDir build"
